Feature/merge readtask task (#64)
* impl Deserialize + Display for signal

* use function instead of static value

* order enum Relaunch

* use serde::default for ReadTask

* remove unwrap_or

* Deserialize Signal using uppercase name

* rework ReadTask using PathBuf

* remove unused AutoRestart

* add method from task::Task

* merge ReadTask with Task

* correct error
