+++ src/server/task.rs
@@ -1,94 +1,198 @@
+use super::{default, error, relaunch::Relaunch, signal::Signal, watcher::Watcher};
 use libc::{gid_t, mode_t, uid_t};
-use serde::Deserialize;
+use serde::{self, Deserialize, Deserializer};
+use std::collections::BTreeMap;
 use std::convert::TryFrom;
-use std::fs::File;
+use std::fmt;
+use std::fs::{self, File};
 use std::os::unix::process::CommandExt;
 use std::path::PathBuf;
 use std::process::{Child, Command};
 use std::vec::Vec;
 
-use super::{error, reader::ReadTask, relaunch::Relaunch, signal};
+pub type ConfigFile = BTreeMap<String, Task>;
 
-#[derive(Debug, Deserialize)]
-enum AutoRestart {
-    Unexpected,
-    True,
-    False,
-}
-
-#[derive(Debug, PartialEq, Eq)]
-pub struct Task {
-    cmd: Vec<String>,
-    numprocess: u32,
-    autostart: bool,
-    umask: mode_t,
-    workingdir: PathBuf,
-
-    stdout: PathBuf,
-    stderr: PathBuf,
-
-    stopsignal: signal::Signal,
-    stopdelay: u32,
-
-    retry: u32,
-
-    successdelay: u32,
-
-    expected_exit_codes: Vec<i32>,
-
-    restart: Relaunch,
-
-    env: Vec<String>,
-    uid: Option<uid_t>,
-    gid: Option<gid_t>,
-}
-
-impl TryFrom<&ReadTask> for Task {
+impl TryFrom<&Watcher> for ConfigFile {
     type Error = error::Taskmaster;
 
-    fn try_from(readtask: &ReadTask) -> Result<Self, Self::Error> {
-        Ok(Self {
-            cmd: readtask
+    fn try_from(watcher: &Watcher) -> Result<Self, error::Taskmaster> {
+        let content: String = match fs::read_to_string(&watcher.path) {
+            Ok(c) => c,
+            Err(e) => return Err(error::Taskmaster::ReadFile(e)),
+        };
+        let ext = watcher.path.extension().and_then(std::ffi::OsStr::to_str);
+
+        match ext {
+            Some("yml") | Some("yaml") => {
+                log::info!("try parsing in YAML format");
+                match serde_yaml::from_str(&content) {
+                    Ok(c) => Ok(c),
+                    Err(e) => Err(error::Taskmaster::ParseYaml(e)),
+                }
+            }
+            Some("toml") => {
+                log::info!("try parsing in TOML format");
+                match toml::from_str(&content) {
+                    Ok(c) => Ok(c),
+                    Err(e) => Err(error::Taskmaster::ParseToml(e)),
+                }
+            }
+            _ => {
+                log::error!("not handler for ext {:?}", ext);
+                Err(error::Taskmaster::Cli)
+            }
+        }
+    }
+}
+
+#[derive(Deserialize)]
+struct TaskPartial {
+    pub cmd: String,
+
+    #[serde(default = "default::autostart")]
+    pub autostart: bool,
+
+    #[serde(default = "default::numprocess")]
+    pub numprocess: u32,
+
+    #[serde(default = "default::umask")]
+    pub umask: mode_t,
+
+    #[serde(default = "default::workdir")]
+    pub workingdir: PathBuf,
+
+    #[serde(default = "default::stop_signal")]
+    pub stopsignal: Signal,
+
+    #[serde(default = "default::stop_delay")]
+    pub stopdelay: u32,
+
+    #[serde(default = "default::stdout")]
+    pub stdout: PathBuf,
+
+    #[serde(default = "default::stderr")]
+    pub stderr: PathBuf,
+
+    #[serde(default = "default::retry")]
+    pub retry: u32,
+
+    #[serde(default = "default::success_delay")]
+    pub successdelay: u32,
+
+    #[serde(default = "default::exit_codes")]
+    pub exitcodes: Vec<i32>,
+
+    #[serde(default = "default::relaunch_mode")]
+    pub restart: Relaunch,
+
+    #[serde(default = "default::env")]
+    pub env: Vec<String>,
+
+    pub gid: Option<gid_t>,
+    pub uid: Option<uid_t>,
+}
+
+#[derive(Debug, Eq, PartialEq, Clone)]
+pub struct Task {
+    cmd: String,
+    args: Vec<String>,
+    pub autostart: bool,
+    numprocess: u32,
+    umask: mode_t,
+    workingdir: PathBuf,
+    stopsignal: Signal,
+    stopdelay: u32,
+    stdout: PathBuf,
+    stderr: PathBuf,
+    retry: u32,
+    successdelay: u32,
+    exitcodes: Vec<i32>,
+    restart: Relaunch,
+    env: Vec<String>,
+    gid: Option<gid_t>,
+    uid: Option<uid_t>,
+}
+
+impl<'de> Deserialize<'de> for Task {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: Deserializer<'de>,
+    {
+        let partial = TaskPartial::deserialize(deserializer)?;
+        Ok(Task::from(partial))
+    }
+}
+
+impl From<TaskPartial> for Task {
+    fn from(partial: TaskPartial) -> Self {
+        Self {
+            cmd: partial.cmd.clone(),
+            args: partial
                 .cmd
                 .split(' ')
                 .map(std::string::ToString::to_string)
                 .collect(),
-            numprocess: readtask.numprocess,
-            autostart: readtask.autostart,
-            umask: readtask.umask,
+            autostart: partial.autostart,
+            numprocess: partial.numprocess,
+            umask: partial.umask,
+            workingdir: partial.workingdir,
+            stopsignal: partial.stopsignal,
+            stopdelay: partial.stopdelay,
+            stdout: partial.stdout,
+            stderr: partial.stderr,
+            retry: partial.retry,
+            successdelay: partial.successdelay,
+            exitcodes: partial.exitcodes,
+            restart: partial.restart,
+            env: partial.env,
+            gid: partial.gid,
+            uid: partial.uid,
+        }
+    }
+}
 
-            retry: readtask.retry,
+impl fmt::Display for Task {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(
+            f,
+            "Command: {}\nNumber of processes: {}\nAutostart: {}\nUmask: {:#05o}\nWorking Directory: {:?}\nStdout: {:?}\nStderr: {:?}\nStop signal: {}\nStop delay: {}\nretry: {}\nSuccess Delay: {}\nExit Codes: {:?}\nRestart: {}\nEnv: {:?}\nPermission: uid: {:?}, gid: {:?}",
+            self.cmd,
+            self.numprocess,
+            self.autostart,
+            self.umask,
+            self.workingdir,
 
-            successdelay: readtask.successdelay,
+            self.stdout,
+            self.stderr,
 
-            uid: readtask.uid,
-            gid: readtask.gid,
+            self.stopsignal,
+            self.stopdelay,
 
-            env: readtask.env.clone(),
+            self.retry,
 
-            restart: readtask.restart.clone(),
+            self.successdelay,
 
-            expected_exit_codes: readtask.exitcodes.clone(),
+            self.exitcodes,
 
-            stopsignal: readtask.stopsignal.clone(),
-            stopdelay: readtask.stopdelay,
+            self.restart,
 
-            workingdir: PathBuf::from(&readtask.workingdir),
-            stdout: PathBuf::from(&readtask.stdout),
-            stderr: PathBuf::from(&readtask.stderr),
-        })
+            self.env,
+
+            self.uid,
+            self.gid,
+        )
     }
 }
 
 impl Task {
     pub fn run(&self) -> Vec<Child> {
         let mut jobs = Vec::new();
-        let mut command = Command::new(&self.cmd[0]);
+        let mut command = Command::new(&self.args[0]);
         let stdout = File::create(self.stdout.as_path()).unwrap();
         let stderr = File::create(self.stderr.as_path()).unwrap();
         self.setup_command(&mut command);
-        if self.cmd.len() > 1 {
-            command.args(&self.cmd[1..]);
+        if self.args.len() > 1 {
+            command.args(&self.args[1..]);
         }
         command.current_dir(self.workingdir.as_path());
         command.stdout(stdout);
@@ -119,9 +223,7 @@ impl Task {
 
     pub fn check_exit_status(&self, status: std::process::ExitStatus) -> bool {
         status.code().map_or(false, |exitcode| {
-            self.expected_exit_codes
-                .iter()
-                .any(|&code| code != exitcode)
+            self.exitcodes.iter().any(|&code| code != exitcode)
         })
     }
 }
